Move index.js files into src dirs, move root index.js to packages/knockout
