--- packages/tko.bind/index.js
@@ -1,6 +1,0 @@
-
-export * from './src/bindingContext';
-export * from './src/applyBindings';
-export * from './src/arrayToDomNodeChildren';
-
-export { BindingHandler, AsyncBindingHandler } from './src/BindingHandler';
