+++ packages/tko.bind/src/index.js
@@ -1,0 +1,6 @@
+
+export * from './bindingContext';
+export * from './applyBindings';
+export * from './arrayToDomNodeChildren';
+
+export { BindingHandler, AsyncBindingHandler } from './BindingHandler';
