--- packages/tko.binding.component/index.js
@@ -1,3 +1,0 @@
-
-import component from './src/componentBinding'
-export var bindings = { component }
