+++ packages/tko.binding.component/package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "description": "component: binding for web components",
   "main": "dist/tko.binding.component.js",
-  "module": "index.js",
+  "module": "src/index.js",
   "license": "MIT",
   "karma": { "frameworks": ["jasmine"]}
 }
