+++ packages/tko.binding.component/src/index.js
@@ -1,0 +1,3 @@
+
+import component from './componentBinding'
+export var bindings = { component }
