+++ packages/tko.binding.foreach/package.json
@@ -3,7 +3,7 @@
   "version": "4.0.0-alpha2",
   "description": "Knockout Foreach Binding",
   "main": "dist/tko.binding.foreach.js",
-  "module": "index.js",
+  "module": "src/index.js",
   "scripts": {
     "test": "gulp test --once"
   },
