+++ packages/tko.computed/package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "description": "TKO Computed Observables",
   "main": "dist/tko.computed.js",
-  "module": "index.js",
+  "module": "src/index.js",
   "dependencies": {
     "tko.utils": "knockout/tko.utils",
     "tko.observable": "knockout/tko.observable"
