+++ packages/tko.filter.punches/package.json
@@ -3,7 +3,7 @@
   "version": "0.2.0",
   "description": "TKO filters from knockout punches",
   "main": "dist/tko.filter.punches.js",
-  "module": "index.js",
+  "module": "src/index.js",
   "files": [
     "dist/",
     "src/",
