--- packages/tko.lifecycle/index.js
@@ -1,3 +1,0 @@
-
-
-export {default as LifeCycle} from './src/LifeCycle'
