+++ packages/tko.lifecycle/src/index.js
@@ -1,0 +1,3 @@
+
+
+export {default as LifeCycle} from './LifeCycle'
