--- packages/tko.observable/index.js
@@ -1,21 +1,0 @@
-//
-// Observables.
-// ---
-//
-// The following are added to the root `[t]ko` object.
-//
-
-export {
-    observable, isObservable, unwrap, peek,
-    isWriteableObservable, isWritableObservable
-} from './src/observable';
-export { isSubscribable, subscribable } from './src/subscribable';
-export { observableArray } from './src/observableArray';
-export { trackArrayChanges, arrayChangeEventName } from './src/observableArray.changeTracking';
-export { toJS, toJSON } from './src/mappingHelpers';
-export { deferUpdates } from './src/defer.js';
-
-export { valuesArePrimitiveAndEqual, applyExtenders, extenders } from './src/extenders';
-import * as dependencyDetection from './src/dependencyDetection';
-
-export { dependencyDetection };
