+++ packages/tko.observable/package.json
@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "description": "TKO subscribables and observables",
   "main": "dist/tko.observable.js",
-  "module": "index.js",
+  "module": "src/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
