+++ packages/tko.observable/src/index.js
@@ -1,0 +1,21 @@
+//
+// Observables.
+// ---
+//
+// The following are added to the root `[t]ko` object.
+//
+
+export {
+    observable, isObservable, unwrap, peek,
+    isWriteableObservable, isWritableObservable
+} from './observable';
+export { isSubscribable, subscribable } from './subscribable';
+export { observableArray } from './observableArray';
+export { trackArrayChanges, arrayChangeEventName } from './observableArray.changeTracking';
+export { toJS, toJSON } from './mappingHelpers';
+export { deferUpdates } from './defer.js';
+
+export { valuesArePrimitiveAndEqual, applyExtenders, extenders } from './extenders';
+import * as dependencyDetection from './dependencyDetection';
+
+export { dependencyDetection };
