--- packages/tko.provider.attr/index.js
@@ -1,1 +1,0 @@
-export {default as AttributeProvider} from './src/AttributeProvider.js'
