+++ packages/tko.provider.attr/package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "description": "Link HTML attributes (e.g. ko-handler-name) to binding handlers",
   "main": "dist/tko.provider.attr.js",
-  "module": "index.js",
+  "module": "src/index.js",
   "license": "MIT",
   "karma": {"frameworks":["mocha","chai"]}
 }
