+++ packages/tko.provider.attr/src/index.js
@@ -1,0 +1,1 @@
+export {default as AttributeProvider} from './AttributeProvider.js'
