--- packages/tko.provider.bindingString/index.js
@@ -1,1 +1,0 @@
-export {default as BindingStringProvider} from './src/BindingStringProvider'
