+++ packages/tko.provider.bindingString/src/index.js
@@ -1,0 +1,1 @@
+export {default as BindingStringProvider} from './BindingStringProvider'
