--- packages/tko.provider.component/index.js
@@ -1,1 +1,0 @@
-export {default as ComponentProvider} from './src/ComponentProvider'
