+++ packages/tko.provider.component/src/index.js
@@ -1,0 +1,1 @@
+export {default as ComponentProvider} from './ComponentProvider'
