--- packages/tko.provider.databind/index.js
@@ -1,1 +1,0 @@
-export {default as DataBindProvider} from './src/DataBindProvider'
