+++ packages/tko.provider.databind/src/index.js
@@ -1,0 +1,1 @@
+export {default as DataBindProvider} from './DataBindProvider'
