--- packages/tko.provider.multi/index.js
@@ -1,2 +1,0 @@
-
-export {default as MultiProvider} from './src/MultiProvider'
