+++ packages/tko.provider.multi/src/index.js
@@ -1,0 +1,2 @@
+
+export {default as MultiProvider} from './MultiProvider'
