--- packages/tko.provider.virtual/index.js
@@ -1,1 +1,0 @@
-export {default as VirtualProvider} from './src/VirtualProvider'
