+++ packages/tko.provider.virtual/package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "description": "Binding provider for <!-- ko handler: ... --> virtual elements",
   "main": "dist/tko.provider.virtual.js",
-  "module": "index.js",
+  "module": "src/index.js",
   "author": "Knockout Team",
   "license": "MIT",
   "karma": { "frameworks": ["mocha", "chai"]}
