--- packages/tko.provider/index.js
@@ -1,2 +1,0 @@
-
-export { default as Provider } from './src/Provider'
