+++ packages/tko.provider/src/index.js
@@ -1,0 +1,2 @@
+
+export { default as Provider } from './Provider'
