--- packages/tko.utils.parser/index.js
@@ -1,7 +1,0 @@
-
-export {default as Parser} from './src/Parser'
-export {default as Identifier} from './src/Identifier'
-export {default as Arguments} from './src/Arguments'
-export {default as Node} from './src/Node'
-
-export {default as parseObjectLiteral} from './src/preparse'
