--- packages/tko.utils/index.js
@@ -1,34 +1,0 @@
-/*
-  tko.util
-  ===
-
-
-*/
-
-export * from './src/array.js';
-export * from './src/async.js';
-export * from './src/error.js';
-export * from './src/ie.js';
-export * from './src/object.js';
-export * from './src/proto.js';
-export * from './src/string.js';
-export * from './src/symbol.js';
-export * from './src/css.js';
-export { jQuerySetInstance } from './src/jquery.js';
-export { default as options } from './src/options.js';
-
-// DOM;
-export * from './src/dom/event.js';
-export * from './src/dom/info.js';
-export * from './src/dom/manipulation.js';
-export * from './src/dom/fixes.js';
-export * from './src/dom/html.js';
-export * from './src/dom/disposal.js';
-
-// Sub-Modules;
-import * as memoization from './src/memoization';
-import * as tasks from './src/tasks.js';
-import * as virtualElements from './src/dom/virtualElements.js';
-import * as domData from './src/dom/data.js';
-
-export {tasks, virtualElements, domData, memoization};
