+++ packages/tko.utils/package.json
@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "description": "TKO Utilities",
   "main": "dist/tko.utils.js",
-  "module": "index.js",
+  "module": "src/index.js",
   "files": [
     "dist/",
     "src/",
